# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sarangi DMS API
  version: 1.0.0-rc2
  description: Backend API for the Sarangi Dental Management System.
  contact:
    name: Sarangi DMS Support
    email: support@example.com
servers:
  - url: https://api.dev.example.com
    description: Dev
  - url: https://api.staging.example.com
    description: Staging
  - url: https://api.example.com
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /patients:
    get:
      tags: [Patients]
      summary: Search patients
      description: >
        Search patients by name or phone-like query. Soft-deleted patients are excluded.
      operationId: listPatients
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: false
          description: Free-text name or phone-like string.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          required: false
          description: Maximum number of patients to return.
      responses:
        '200':
          description: Patients list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  nextCursor:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Patients]
      summary: Create patient
      description: >
        Creates a new patient. Enforces uniqueness on the composite of normalized phone + name.
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate patient (same normalized phone + name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /patients/{patientId}:
    get:
      tags: [Patients]
      summary: Get patient by id
      description: >
        Returns a single patient. Soft-deleted patients are treated as not found.
      operationId: getPatientById
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or soft-deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Patients]
      summary: Update patient
      description: >
        Partially updates a patient. Enforces composite phone+name uniqueness on change.
      operationId: updatePatient
      parameters:
        - $ref: '#/components/parameters/PatientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate patient (same normalized phone + name)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /patients/{patientId}/visits:
    get:
      tags: [Patients, Visits]
      summary: List visits for a patient
      description: Lists all visits associated with a given patient.
      operationId: listPatientVisits
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Visit list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Visit'
        '400':
          description: Invalid patient id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /medicines:
    get:
      tags: [Medicines]
      summary: Search medicines
      description: >
        Typeahead search for medicine presets based on display name.
      operationId: searchMedicines
      parameters:
        - in: query
          name: query
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Medicine presets
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MedicineTypeaheadItem'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /medicines/quick-add:
    post:
      tags: [Medicines]
      summary: Quick-add a medicine preset
      description: >
        Quickly creates a new medicine preset sourced from inline doctor entry.
      operationId: quickAddMedicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuickAddMedicineInput'
      responses:
        '201':
          description: Medicine preset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicinePreset'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rx-presets:
    get:
      tags: [Prescriptions]
      summary: Search prescription presets
      description: >
        Typeahead search over prescription presets.
      operationId: searchPrescriptionPresets
      parameters:
        - in: query
          name: query
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Prescription presets
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PrescriptionPreset'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/rx-presets:
    post:
      tags: [Admin]
      summary: Create prescription preset
      description: ADMIN-only. Creates a new prescription preset.
      operationId: adminCreateRxPreset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateRxPresetRequest'
      responses:
        '201':
          description: Prescription preset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionPreset'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/rx-presets/{id}:
    patch:
      tags: [Admin]
      summary: Update prescription preset
      description: ADMIN-only. Partially updates a prescription preset.
      operationId: adminUpdateRxPreset
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateRxPresetRequest'
      responses:
        '200':
          description: Prescription preset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrescriptionPreset'
        '400':
          description: Invalid id or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Preset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PatientId:
      in: path
      name: patientId
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
          description: Human-readable error message.
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Optional field-level validation errors.
        traceId:
          type: string
...
    PatientCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
    Patient:
      type: object
      required: [patientId, name, createdAt, updatedAt]
      properties:
        patientId:
          type: string
        name:
          type: string
        phone:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, OTHER, UNKNOWN]
        createdAt:
          type: integer
        updatedAt:
          type: integer
        isDeleted:
          type: boolean
        deletedAt:
          type: integer
    PatientUpdate:
      type: object
      description: Partial update for patient. At least one field is required.
      properties:
        name:
          type: string
        phone:
          type: string
        dob:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
...
    Visit:
      type: object
      required: [visitId, patientId, doctorId, reason, status, visitDate, createdAt, updatedAt]
      properties:
        visitId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [QUEUED, IN_PROGRESS, DONE]
        visitDate:
          type: string
          format: date
        createdAt:
          type: integer
        updatedAt:
          type: integer
        billingAmount:
          type: number
...
    RxLine:
      type: object
      required: [medicine, dose, frequency, duration]
      properties:
        medicine:
          type: string
        dose:
          type: string
        frequency:
          type: string
          enum: [QD, BID, TID, QID, HS, PRN]
        duration:
          type: integer
          minimum: 1
          maximum: 365
        sig:
          type: string
        timing:
          type: string
          enum: [BEFORE_MEAL, AFTER_MEAL, ANY]
        notes:
          type: string
...
    MedicinePreset:
      type: object
      required: [id, displayName, normalizedName, createdAt, createdByUserId, source, verified]
      properties:
        id:
          type: string
        displayName:
          type: string
        normalizedName:
          type: string
        defaultDose:
          type: string
        defaultFrequency:
          type: string
        defaultDuration:
          type: integer
        form:
          type: string
          enum: [TABLET, CAPSULE, SYRUP, INJECTION, OINTMENT, GEL, MOUTHWASH, OTHER]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: integer
        createdByUserId:
          type: string
        source:
          type: string
          enum: [ADMIN_IMPORT, INLINE_DOCTOR]
        verified:
          type: boolean
    MedicineTypeaheadItem:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: string
        displayName:
          type: string
        defaultFrequency:
          type: string
        defaultDuration:
          type: integer
        form:
          type: string
          enum: [TABLET, CAPSULE, SYRUP, INJECTION, OINTMENT, GEL, MOUTHWASH, OTHER]
    QuickAddMedicineInput:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
        defaultDose:
          type: string
        defaultFrequency:
          type: string
        defaultDuration:
          type: integer
          minimum: 1
          maximum: 365
        form:
          type: string
          enum: [TABLET, CAPSULE, SYRUP, INJECTION, OINTMENT, GEL, MOUTHWASH, OTHER]
    PrescriptionPreset:
      type: object
      required: [id, name, lines, createdByUserId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RxLine'
        tags:
          type: array
          items:
            type: string
        createdByUserId:
          type: string
        createdAt:
          type: integer
    AdminCreateRxPresetRequest:
      type: object
      required: [name, lines]
      properties:
        name:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RxLine'
        tags:
          type: array
          items:
            type: string
    AdminUpdateRxPresetRequest:
      type: object
      properties:
        name:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/RxLine'
        tags:
          type: array
          items:
            type: string
...